Add BorderRadius support for UI Nodes
This is a first implementation that's very simple and naive.
I'd like to add support for a few other features like border color
and drop shadow. Having all of that data in the vertex buffer will
become overwhelming pretty quickly, so I will look into other ways
to implement this.