Use uniform buffer for new UI node features
This is a WIP and is not working as of now. It seems like I may
have some issues with the length of UiUniform/UiUniformEntry
because I can see that only one of my corners are rounded,
and nodes index 1+ from each batch clearly have data that's
all wrong, because they don't even render.